rework submit and main, found bug with run
