+++ SchedulingComponents/SchedulingThread.java
@@ -1,6 +1,5 @@
 package SchedulingComponents;
 import BatchJobsComponents.BatchJob;
-import Ux.UIHelper;
 import java.util.Scanner;
 import java.util.concurrent.BlockingQueue;
 
@@ -53,17 +52,21 @@ public class SchedulingThread extends Thread {
     /**
      * Run the scheduling thread
      */
-    public void run() {
+    public void run(String userInput) {
         SchedulingPolicy schedulingPolicy = new SchedulingPolicy();
         Scanner sc = new Scanner(System.in);
         //Add jobs to the queue
-        String userInput = sc.nextLine();
-        String[] jobDetails = userInput.split(" ");
+        // String userInput = sc.nextLine();
+        String[] jobDetails = userInput.split("\\s+");
         if(!userInput.equals("exit")) {
             if(jobDetails.length == 3) {
                 BatchJob job = new BatchJob(jobDetails[0], Integer.parseInt(jobDetails[1]), Integer.parseInt(jobDetails[2]));
                 try {
                     jobQueue.put(job);
+                    System.out.println("Job " + job.getJobName() + " was submitted");
+                    System.out.println("Total number of jobs in the queue: " + jobQueue.size());
+                    System.out.println("Expected waiting time: ");
+                    System.out.println("Scheduling Policy" + getSchedulingPolicy());
                 } catch (InterruptedException e) {
                     Thread.currentThread().interrupt();
                 }
@@ -87,21 +90,5 @@ public class SchedulingThread extends Thread {
             this.setJobQueue(priority);
         }
     }
-
-    public static void main(String[] args) {
-        UIHelper helper = new UIHelper();
-        System.out.println("Welcome to CSU BatchJob Please Enter a Command");
-        System.out.println("Enter help for more options");
-
-        Scanner keyboard = new Scanner(System.in);
-        String command = keyboard.nextLine();
-
-        while(!command.equals("quit")){
-            helper.commandFunction(command);
-            command = keyboard.nextLine();
-        }
-        keyboard.close();
-
-    };
     
 }
